Added initialize for vault and testing
Added a state for initializing a newly installed Vault server. It can
optionally be unsealed automatically as well. If a list of PGP keys or
Keybase usernames are provided then the sealing keys will be regenerated
and backed up after the vault is unsealed.

Added Testinfra tests to validate the state of the configured system.