Moved cached read function to execution module
In order to allow for leveraging cached leases from jinja templates I moved the cached_read function from the renderer module to the execution module. This will allow us to use those leases in cases such as where there is a full connection string for postgres that needs to use the username and password in the same pillar value.